Ensure migrations skip existing structures

+++ db/migrate/20221019171906_community_add_default_period_to_leaderboards.rb
@@ -1,10 +1,12 @@
 # frozen_string_literal: true
 class CommunityAddDefaultPeriodToLeaderboards < ActiveRecord::Migration[6.1]
   def up
-    add_column :gamification_leaderboards, :default_period, :integer, default: 0
+    unless column_exists?(:gamification_leaderboards, :default_period)
+      add_column :gamification_leaderboards, :default_period, :integer, default: 0
+    end
   end
 
   def down
-    remove_column :gamification_leaderboards, :default_period
+    remove_column :gamification_leaderboards, :default_period if column_exists?(:gamification_leaderboards, :default_period)
   end
 end
